Add Eureka and Config Server for Chat-Agent and use stable 1.0.0-M2 package version (#62)
## Purpose

+ Add Eureka and Config Server
+ Use the stable 1.0.0-M2 spring-ai-azure-openai package version

## Does this introduce a breaking change?
<!-- Mark one with an "x". -->
```
[ ] Yes
[ ] No
```

## Pull Request Type
What kind of change does this Pull Request introduce?

<!-- Please check the one that applies to this PR using "x". -->
```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Documentation content changes
[ ] Other... Please describe:
```

## How to Test
*  Get the code

```
git clone [repo-address]
cd [repo-name]
git checkout [branch-name]
npm install
```

* Test the code
<!-- Add steps to run the tests suite and/or manually test -->
```
```

## What to Check
Verify that the following are valid
* ...

## Other Information
<!-- Add any other helpful information that may be needed here. -->

---------

+++ src/spring-petclinic-chat-agent/src/main/java/org/springframework/samples/petclinic/agent/chat/ChatOptionsProperties.java
@@ -13,6 +13,6 @@ public class ChatOptionsProperties {
 
     String deploymentName;
 
-    Double temperature;
+    Float temperature;
 
 }
